// app/types.ts
export interface SloganResponseType {
  data: {
    input: string;
    num_sequences: number;
    language: string;
    model: string;
  };

  result: {
    id: number;
    ms: number;
    device: string;
    predictions: string[];
  };

  datetime: number;
}

export interface LogoResponseType {
  data: {
    brand_personality: string;
    brand_name: string;
    category: string;
    industry: string;
    num_images: number;
    num_inference_steps: number;
    lan: string;
    prompt: string;
  };

  result: {
    id: number;
    ms: number;
    predictions: string[];
    recommend: Recommend[];
  };

  datetime: number;
}

export interface Recommend {
  order: number;
  score: number;
}
// ...
export interface TestLogoInput {
  category: string;
  positive_prompt: string;
  negative_prompt: string;
  num_images: string;
  num_inference_steps: string;
  pixel_resolution: string;
  compression_format: string;
  type: string;
}